move all queries to mediarepository
